refactor(sdk): improve daemon-mcp architecture (P1/P2 fixes)
P1 fixes:
- Split types.ts into types.ts (interfaces), sse.ts (SSE lifecycle),
  helpers.ts (handler/resolveSessionId/daemonFetch) for separation of concerns
- Add fileStat/dirList/glob methods to DaemonClient, removing raw
  daemonFetch usage from workspaceRead tools
- Move session_set_model and session_context from agent.ts to session.ts
  for naming consistency
- Add error logging with stack traces in handler() wrapper

P2 fixes:
- Remove unused exports from formatters.ts (formatToolResult,
  formatTextResult, mergeToolResults, isValidContentBlock)
- Fix copyright year to 2026 in tool.ts and createSdkMcpServer.ts
