+++ packages/sdk-typescript/src/daemon-mcp/createSdkMcpServer.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2025 Qwen Team
+ * Copyright 2026 Qwen Team
  * SPDX-License-Identifier: Apache-2.0
  */
