+++ packages/sdk-typescript/src/daemon-mcp/formatters.ts
@@ -1,139 +1,16 @@
 /**
- * Tool result formatting utilities for MCP responses
- *
- * Converts various output types to MCP content blocks.
+ * @license
+ * Copyright 2026 Qwen Team
+ * SPDX-License-Identifier: Apache-2.0
  */
 
-export type McpContentBlock =
-  | { type: 'text'; text: string }
-  | { type: 'image'; data: string; mimeType: string }
-  | { type: 'resource'; uri: string; mimeType?: string; text?: string };
+/**
+ * Tool result formatting utilities for MCP responses.
+ */
 
 export interface ToolResult {
-  content: McpContentBlock[];
+  content: Array<{ type: 'text'; text: string }>;
   isError?: boolean;
-}
-
-export function formatToolResult(result: unknown): ToolResult {
-  // Handle Error objects
-  if (result instanceof Error) {
-    return {
-      content: [
-        {
-          type: 'text',
-          text: result.message || 'Unknown error',
-        },
-      ],
-      isError: true,
-    };
-  }
-
-  // Handle null/undefined
-  if (result === null || result === undefined) {
-    return {
-      content: [
-        {
-          type: 'text',
-          text: '',
-        },
-      ],
-    };
-  }
-
-  // Handle string
-  if (typeof result === 'string') {
-    return {
-      content: [
-        {
-          type: 'text',
-          text: result,
-        },
-      ],
-    };
-  }
-
-  // Handle number
-  if (typeof result === 'number') {
-    return {
-      content: [
-        {
-          type: 'text',
-          text: String(result),
-        },
-      ],
-    };
-  }
-
-  // Handle boolean
-  if (typeof result === 'boolean') {
-    return {
-      content: [
-        {
-          type: 'text',
-          text: String(result),
-        },
-      ],
-    };
-  }
-
-  // Handle object (including arrays)
-  if (typeof result === 'object') {
-    try {
-      return {
-        content: [
-          {
-            type: 'text',
-            text: JSON.stringify(result, null, 2),
-          },
-        ],
-      };
-    } catch {
-      // JSON.stringify failed
-      return {
-        content: [
-          {
-            type: 'text',
-            text: String(result),
-          },
-        ],
-      };
-    }
-  }
-
-  // Fallback: convert to string
-  return {
-    content: [
-      {
-        type: 'text',
-        text: String(result),
-      },
-    ],
-  };
-}
-
-export function formatToolError(error: Error | string): ToolResult {
-  const message = error instanceof Error ? error.message : error;
-
-  return {
-    content: [
-      {
-        type: 'text',
-        text: message,
-      },
-    ],
-    isError: true,
-  };
-}
-
-export function formatTextResult(text: string): ToolResult {
-  return {
-    content: [
-      {
-        type: 'text',
-        text,
-      },
-    ],
-  };
 }
 
 export function formatJsonResult(data: unknown): ToolResult {
@@ -147,48 +24,10 @@ export function formatJsonResult(data: unknown): ToolResult {
   };
 }
 
-export function mergeToolResults(results: ToolResult[]): ToolResult {
-  const mergedContent: McpContentBlock[] = [];
-  let hasError = false;
-
-  for (const result of results) {
-    mergedContent.push(...result.content);
-    if (result.isError) {
-      hasError = true;
-    }
-  }
-
+export function formatToolError(error: Error | string): ToolResult {
+  const message = error instanceof Error ? error.message : error;
   return {
-    content: mergedContent,
-    isError: hasError,
+    content: [{ type: 'text', text: message }],
+    isError: true,
   };
-}
-
-export function isValidContentBlock(block: unknown): block is McpContentBlock {
-  if (!block || typeof block !== 'object') {
-    return false;
-  }
-
-  const blockObj = block as Record<string, unknown>;
-
-  if (!blockObj.type || typeof blockObj.type !== 'string') {
-    return false;
-  }
-
-  switch (blockObj.type) {
-    case 'text':
-      return typeof blockObj.text === 'string';
-
-    case 'image':
-      return (
-        typeof blockObj.data === 'string' &&
-        typeof blockObj.mimeType === 'string'
-      );
-
-    case 'resource':
-      return typeof blockObj.uri === 'string';
-
-    default:
-      return false;
-  }
 }
