+++ packages/sdk-typescript/src/daemon-mcp/serve-bridge/sse.ts
@@ -1,0 +1,91 @@
+/**
+ * @license
+ * Copyright 2026 Qwen Team
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * Persistent SSE connection lifecycle management.
+ */
+
+import type {
+  BridgeState,
+  PromptCollector,
+  SessionEventStream,
+} from './types.js';
+
+/**
+ * Create a new PromptCollector that resolves when called.
+ */
+export function createPromptCollector(): PromptCollector {
+  let resolve!: () => void;
+  const promise = new Promise<void>((r) => {
+    resolve = r;
+  });
+  return { texts: [], resolve, promise };
+}
+
+/**
+ * Start a persistent SSE subscription for a session.
+ * Collects agent_message_chunk events into the active PromptCollector.
+ */
+export function startEventStream(state: BridgeState, sessionId: string): void {
+  // Don't create duplicate streams
+  if (state.eventStreams.has(sessionId)) return;
+
+  const abortCtrl = new AbortController();
+  const stream: SessionEventStream = {
+    sessionId,
+    abortCtrl,
+    activeCollector: null,
+  };
+  state.eventStreams.set(sessionId, stream);
+
+  // Start consuming SSE in the background (fire-and-forget)
+  (async () => {
+    try {
+      for await (const event of state.client.subscribeEvents(sessionId, {
+        signal: abortCtrl.signal,
+      })) {
+        const data = event.data as Record<string, unknown> | undefined;
+        if (!data) continue;
+        const update = data['update'] as Record<string, unknown> | undefined;
+        if (!update) continue;
+        if (update['sessionUpdate'] === 'agent_message_chunk') {
+          const content = update['content'] as
+            | Record<string, unknown>
+            | undefined;
+          if (!content) continue;
+          const collector = stream.activeCollector;
+          if (collector) {
+            const text = content['text'];
+            if (typeof text === 'string' && text) {
+              collector.texts.push(text);
+            }
+            // _meta signals end of the current message
+            if ('_meta' in content) {
+              collector.resolve();
+            }
+          }
+        }
+      }
+    } catch {
+      // SSE disconnected or aborted — expected on session close
+    } finally {
+      state.eventStreams.delete(sessionId);
+    }
+  })();
+}
+
+/**
+ * Stop the persistent SSE subscription for a session.
+ */
+export function stopEventStream(state: BridgeState, sessionId: string): void {
+  const stream = state.eventStreams.get(sessionId);
+  if (stream) {
+    stream.abortCtrl.abort();
+    // Resolve any pending collector so prompt doesn't hang
+    stream.activeCollector?.resolve();
+    state.eventStreams.delete(sessionId);
+  }
+}
